Fix stageout path handling.
Set the stageout storage to be the configured stageout directory plus the
label of the sample.  This makes the path shorter, as arbitrary working
directories can be specified.  Also, a working directory with a leading /
broke the code, leading to output files being stored in the working
directory instead of in the stageout directory.