Run the python job wrapper on non-cmsRun jobs.
It is meant to be an intermediate measure so that we can
have access to non-wq stagein and stageout for non-cmsRun jobs
while we are working out the details of #85; it is not meant
to address any of the issues discussed there. Note that I chose to
use xml.dom.minidom for writing the dummy FJR; there are native
CMSSW tools for this but I found them obtuse to deal with-- they
may be worth revisiting.