Handle task results from WQ. Fixes #80.
I believe this is the correct subset of WQ results for which we
want the WQ result to trump whatever exit code we're returning. For
missing files (wq.WORK_QUEUE_RESULT_{INPUT|OUTPUT}_MISSING, let's
handle it on our side (trust the wrapper to return a sensible exit
code), as it gives us more control.