Move things around a bit.
Put the JobHandler into its own file and rename it TaskHandler, in
agreement with #133.  At the same time, factor out all the dicts used to
keep track of processing related things into a Workflow class.

In the end, the JobSource.obtain method should be legible from a logical
viewpoint, with adjustments done by the Workflow and TaskHandler classes
for processing and task-specific changes, respectively.