Allow to remotely change Lobster configuration.
Fixes #276, closes #28.  When running a Lobster process, `lobster
reconfigure <setting> <value> <config>` will send a message to the main
process, that will get parsed at one point in the main loop.  If the
message sent is a valid command, Lobster will alter the corresponding
setting.

Relies on a lot more work to make changing settings actually work
(everything that is changeable should refer to the config and not copy
values out of the config).