Remove file-interface yaml example.
I'm not replacing this with a python equivalent: now that we
have better documentation, the purpose of these examples
should be to quickly get people up and running with a simple
working starting point from which to make modifications,
rather than to show all configuration possibilities.
Therefore I don't think we need examples for each different
kind of input dataset-- if that is unclear, we should
clarify it in the docs. Too many examples just makes the
examples directory more confusing.