Factor out repeated code for traversing filelist
As we have several baseclasses here I am not making this a method
so that it can be mixed in without multiple inheritance. Note that
the previous `Dataset.validate()` method is satisfied as long as
the directory is valid, even if it is empty. Now there must be
at least one input file.