Deduplicate units stuck
Under some conditions a unit can satisfy both being failed and being
skipped. In those cases we double-count how many units are stuck (this
error was introduced by me in c98a5bb). This commit reverts back to
calculating the union of stuck and failed rather than the sum. It also
changes the definition of 'stuck' given in workflow summary to be in
line with what we use under-the-hood (so for users, stuck goes from
"parent stuck" to "parent stuck or skipped or failed". I think the
former use is more straightforward for users, but maintaining a separate
definition is too confusing. Fixes #592.