Make commands a lot more customizable.
Remove `pset` and `argument` parameters, and instead parse the `command`
parameter for workflows. This allows to shuffle things around a little and
will be more flexible in the future. The following merge command would now
be possible:

    my_merge.py -o @outputfiles --inputs @inputfiles

While at it, pulling in datasets from DBS and splitting them on files broke
recently, fix it.

Fixes #207.