Try harder to check the files after stageout
Following the patented "Lobster method" don't accept defeat immediately if your attempt to compare the remote file size against the local one fails.  This might be because the transfer failed, but it also might be because your method of checking the file size for the remote file is failing.  Be more resilient!  Check the file size with other methods and only give up on the transfer if the file size is incompatible with all methods.