Address Issues with WQ Transfers of report.json
The `report.json` file initially contains an empty json object which is read into the code and then populated by data in the tasks.  The previous strategy for handling this file was to transfer it to worker using the WQ transfer mechanism with the local name of `report.json` but the remote name of `report.json.in`.  The `wrapper.sh` script would `mv report.json.in report.json` and then `task.py` would read and write the file with that name.

I'm not sure I completely understand, but WQ was having some issues with the fact that this file was an input and an output for the worker and that the local name of the input was the same as the remote and local name for the output.  The new strategy is to keep the template file consistently names `report.json.in` on the master and worker side.  The `wrapper.sh` script doesn't rename it.  The `task.py` script initially reads in `report.json.in` but writes out (and subsequently re-reads) `report.json`.